Use Thermite to build extension
This commit adds Thermite to build and compile the gem extension. Using
Thermite, users of the gem do not need to have a working Rust compiler
in order to install the gem. Instead, Thermite will check for a compiled
library matching the user's environment on GitHub Releases and download
that instead. If there is no matching compiled library available, then
Thermite falls back to compiling the extension locally.

This makes this gem more accessible as users do not need to understand
how to install and manage Rust in order to use the gem.